pyproject.toml
[tool.poetry]
name = "openai_model_tracker"
version = "0.1.0"
description = ""
authors = ["Author Name <author@example.com>"]
readme = "README.md"
packages = [
  { include = "openai_model_tracker" }
]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
tabulate = "^0.9.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
mypy = "^1.11.0"
types-requests = "^2.31.0"
types-tabulate = "^0.9.0"

[tool.poetry.scripts]
openai_model_tracker = "openai_model_tracker.__main__:main"

[tool.black]
line-length = 88

[tool.mypy]
strict = true
disallow_any_unimported = false
disallow_any_decorated = false
disallow_any_explicit = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
